requirements.txt
appdirs==1.4.3
packaging==16.8
pyparsing==2.2.0
requests==2.13.0
six==1.10.0
pyTelegramBotAPI
mysql-connector-python==8.0.29
urllib3